+++ Jetpack/Assets/Scenes/StartScreen.unity
@@ -290,7 +290,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &88663482
 MonoBehaviour:
@@ -335,6 +335,52 @@ MonoBehaviour:
     mipBias: 0
     varianceClampScale: 0.9
     contrastAdaptiveSharpening: 0
+--- !u!1 &416182025
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 416182027}
+  - component: {fileID: 416182026}
+  m_Layer: 0
+  m_Name: LoadDelay
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &416182026
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 416182025}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fa50f6a36763497982f90c6d63d1ab63, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  levelLoader: {fileID: 1924974226}
+  loadDelay: 3
+--- !u!4 &416182027
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 416182025}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 597.2775, y: 258.05878, z: -3.694451}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1575726396
 GameObject:
   m_ObjectHideFlags: 0
@@ -477,7 +523,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!95 &1921836051 stripped
 Animator:
@@ -494,7 +540,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 7071527113645546177, guid: 2ca6d4272adb4944db176ea08f739eae, type: 3}
       propertyPath: m_RootOrder
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7071527113645546177, guid: 2ca6d4272adb4944db176ea08f739eae, type: 3}
       propertyPath: m_LocalPosition.x
@@ -549,6 +595,11 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 2ca6d4272adb4944db176ea08f739eae, type: 3}
+--- !u!1 &1924974226 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 8819095057015350296, guid: 2ca6d4272adb4944db176ea08f739eae, type: 3}
+  m_PrefabInstance: {fileID: 1924974225}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1931403543
 GameObject:
   m_ObjectHideFlags: 0
@@ -646,7 +697,7 @@ RectTransform:
   - {fileID: 66917406}
   - {fileID: 1575726397}
   m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
