+++ Jetpack/Assets/Scripts/Scenes/GameStartDelay.cs
@@ -1,0 +1,10 @@
+﻿namespace Scenes
+{
+    public class GameStartDelay : LoadDelay
+    {
+        protected override void Logic()
+        {
+            this.LevelLoaderScript.LoadStartMenuScene();
+        }
+    }
+}
